+++ CHANGELOG.md
@@ -6,6 +6,43 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
 
+## 2025.04.18
+
+### Added
+
+- Designed test suite for playlist component
+- Designed two different use cases for playlist component
+
+### Updated
+
+- Added removeSong(String name) to secondary interface/implementation
+- Cleaned up doc folder
+
+## 2025.04.09
+
+### Added
+
+- Designed kernel implementation for playlist component
+- Added toString and equals to secondary implementation
+
+## 2025.04.04
+
+### Added
+
+- Designed abstract class for playlist component
+
+### Updated
+
+- Added length() to kernel interface
+- Changed remove() to removeFront()
+- Added randomSong() to secondary interface
+
+## 2025.03.09
+
+### Added
+
+- Designed kernel and enhanced interfaces for playlist component
+
 ## 2025.01.31
 
 ### Added